Add schema for job_types table
We need a separate job_types table for faster query and reduce redundancy.
The job name and stage columns are separated from the jobs table and metadata
like these are stored in the job_types table instead.

This separation is also required in regards of collection of stats of only the
jobs that belong to the official Salsa CI pipeline.

We use the job's name to differentiate between job types. This could be prone
to error, for example when a modified salsa-ci runs with the jobs of same name,
but this could be safely ignored because of its rarity.