Add schema for statistics tables
We are storing the stats in the tables in attempts to reduce excessive load
from our database and for faster responses.

We achieve this by adding daily and hourly stats tables for both job types and pipelines.
This design, in my understanding, provides proper granularity for populating the graphs
on the frontend, which are the ultimate benefactrs of these data.