Implement cron jobs for statistics calculation and insertion
In the current design, hourly and daily stats are collected on-the-go, that is,
we collect hourly  stats for last hour and insert it, and similarly for daily stats as well.

The cron jobs are scheduled after a margin. E.g., 5 mins after the hour ends to ensure
proper collection of stats of previous hour.