Only calculate job type stats for official Salsa CI jobs
Updated stats calculation queries to only include job types with origin 'salsaci'.
This ensures that statistics are only calculated for jobs that belong to the official
Salsa CI pipeline. This fact can be either manually marked or determined by the high
number of jobs for that job type, although, this automation is yet to be implemented.