Fix the bug where pipelines running > 1h weren't accounted for
The pipelines running longer than 1h weren't _missed_ since
we only checked for the last hours and use the started_at field as the
duration filter.

For a fix, using finished_at to query the pipelines in the past
hour would have made sense to reduce the number of database operations,
however, this will severely misrepresent the number of pipelines *starting*
in the respective hours and will not be helpful in the line chart to track
trends after any event such as merger, regression, etc.

So we are settling with checking hourly stats for each hour in the past
$DEFAULT_MAX_DURATION_PIPELINES hours, where DEFAULT_MAX_DURATION_PIPELINES
is the maximum duration any pipeline can possibly run.

The same will be implemented for the jobs in the commit that follows.

Partially addresses #1