Fix the bug where jobs running > 1h weren't accounted for
This commit is a continuation of the previous one, where we fix
the same issue for pipelines. It is split into 2 commits to improve
readability.

The fix involves checking the jobs for last DEFAULT_MAX_DURATION_JOBS
hours and updating/inserting those stats. For more context see the
previous commit.

Closes #1