Ensure timeline covers (start, end] with no gaps for hourly/daily data
Ensure we return data with a consistent timeline (x-axis) with no gaps.

Earlier it sent only the days which had data, with gaps in them. This caused
several rendering issues when mapped to Nivo Radar Chart. Apart from these, the
Line Chart and Radar chart's connecting lines were continuous even when there
was no data in between because the corresponding x-axis values were missing.
Filling empty days with nulls breaks those lines and provide a more accurate
visualization.

We now fill the x-axis consistently according to the duration and populate the
data with null values where we don't have any.